todo/requirements.txt
requests
Flask
opentelemetry-sdk
opentelemetry-distro
opentelemetry-instrumentation
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-flask
opentelemetry-exporter-jaeger
protobuf==3.20.*